requirements.txt
flask==2.3.3
python-dotenv==1.0.0
groq==0.9.0
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
Pillow==10.0.0
matplotlib>=3.5.0
seaborn>=0.11.0